update with ES6/TypeScript support for v3
